refactor(data,viewer): shed the two modules the new spatial types pushed over budget
Adding IfcMarinePart / IfcFacilityPartCommon put two allowlisted modules
past their ratchet budgets. Neither budget is raised; both files shrink.

packages/data/src/types.ts (641 > 628): TYPE_ENUM_TO_STRING was a
hand-written second copy of the IfcTypeEnum member list — 127 rows, every
one mapping a member to a string spelling its own name, with nothing
keeping the two in step. TypeScript's numeric-enum reverse mapping already
answers that question, so IfcTypeEnumToString reads IfcTypeEnum[type]
directly and the table is gone. Verified equivalent by compiling both
versions and comparing their output over every enum member and every id
in -5..399: zero differences, and every member still round-trips through
IfcTypeEnumFromString. 641 -> 508.

apps/viewer/.../treeDataBuilder.ts (1471 > 1466): GROUP_ENTITY_TYPES,
GroupSubFilter, SYSTEM_GROUP_TYPES and groupMatchesSubFilter move to a
sibling groupEntityTypes.ts and are re-exported, so every existing import
keeps resolving. Pure move. 1471 -> 1424.

Allowlist rows LOWERED, none raised: treeDataBuilder.ts 1466 -> 1424,
packages/data/src/types.ts 628 -> 508. ALLOWLIST_DIGEST re-pinned in the
same commit.
